+++ packages/alfa-css/test/value/shape/circle.spec.ts
@@ -1,23 +1,14 @@
-import { Assertions, test } from "@siteimprove/alfa-test";
-
-import { Slice } from "@siteimprove/alfa-slice";
+import { test } from "@siteimprove/alfa-test";
 
 import { Lexer } from "../../../src/syntax/lexer";
 import { Circle } from "../../../src/value/shape/circle";
 
-function parse(t: Assertions, input: string, expected: Circle.JSON) {
-  t.deepEqual(
-    Circle.parse(Slice.of(Lexer.lex(input)))
-      .map(([_, circle]) => circle)
-      .get()
-      .toJSON(),
-    expected,
-    input
-  );
+function parse(input: string) {
+  return Circle.parse(Lexer.lex(input)).map(([_, circle]) => circle.toJSON());
 }
 
 test("parse() parses a circle with just a radius", (t) => {
-  parse(t, "circle(farthest-side)", {
+  t.deepEqual(parse("circle(farthest-side)").get(), {
     type: "basic-shape",
     kind: "circle",
     radius: {
@@ -43,7 +34,7 @@ test("parse() parses a circle with just a radius", (t) => {
 });
 
 test("parse() parses a circle with just a center", (t) => {
-  parse(t, "circle(at left)", {
+  t.deepEqual(parse("circle(at left)").get(), {
     type: "basic-shape",
     kind: "circle",
     radius: {
@@ -73,7 +64,7 @@ test("parse() parses a circle with just a center", (t) => {
 });
 
 test("parse() parses a circle with both radius and center", (t) => {
-  parse(t, "circle(10px at left)", {
+  t.deepEqual(parse("circle(10px at left)").get(), {
     type: "basic-shape",
     kind: "circle",
     radius: {
@@ -104,5 +95,5 @@ test("parse() parses a circle with both radius and center", (t) => {
 });
 
 test("parse() fails if there is a negative radius", (t) => {
-  t.deepEqual(Circle.parse(Slice.of(Lexer.lex("circle(-1px)"))).isErr(), true);
+  t.deepEqual(parse("circle(-1px)").isErr(), true);
 });